+++ tests/utils/keyring_utils.py
@@ -42,41 +42,8 @@ def is_keyring_available():
     Return boolean indicating whether the keyring service is available on
     the local system.
     """
-
-    # Check some obvious cases where no keyring service is available
-
-    backend = keyring.get_keyring()
-    if isinstance(backend, keyring.backends.chainer.ChainerBackend):
-        if not backend.backends:
-            # Chainer backend with empty list of real backends
-            return False
-    if isinstance(backend, keyring.backends.fail.Keyring):
-        return False
-    if isinstance(backend, keyring.backends.null.Keyring):
-        return False
-
-    # In theory, now the keyring service should be available.
-    # We try it out to make really sure.
-
     keyringlib = KeyRingLib()
-    service = keyringlib.keyring_service()
-    username = keyringlib.keyring_username('test/is_keyring_available')
-
-    # Exception to catch when no keyring service is available.
-    # Keyring version 22.0 introduced NoKeyringError and before that used
-    # RuntimeError.
-    try:
-        no_keyring_exception = keyring.errors.NoKeyringError
-    except AttributeError:
-        no_keyring_exception = RuntimeError
-
-    try:
-        keyring.set_password(service, username, 'dummy')
-    except no_keyring_exception:
-        return False
-    else:
-        keyring.delete_password(service, username)
-        return True
+    return keyringlib.is_available()
 
 
 def remove_keyring_item(filepath):